+++ pkg/cloudapi/account/create.go
@@ -23,23 +23,23 @@ type CreateRequest struct {
 
 	// Max size of memory in MB
 	// Required: false
-	MaxMemoryCapacity uint64 `url:"maxMemoryCapacity,omitempty"`
+	MaxMemoryCapacity int64 `url:"maxMemoryCapacity,omitempty"`
 
 	// Max size of aggregated vdisks in GB
 	// Required: false
-	MaxVDiskCapacity uint64 `url:"maxVDiskCapacity,omitempty"`
+	MaxVDiskCapacity int64 `url:"maxVDiskCapacity,omitempty"`
 
 	// Max number of CPU cores
 	// Required: false
-	MaxCPUCapacity uint64 `url:"maxCPUCapacity,omitempty"`
+	MaxCPUCapacity int64 `url:"maxCPUCapacity,omitempty"`
 
 	// Max sent/received network transfer peering
 	// Required: false
-	MaxNetworkPeerTransfer uint64 `url:"maxNetworkPeerTransfer,omitempty"`
+	MaxNetworkPeerTransfer int64 `url:"maxNetworkPeerTransfer,omitempty"`
 
 	// Max number of assigned public IPs
 	// Required: false
-	MaxNumPublicIP uint64 `url:"maxNumPublicIP,omitempty"`
+	MaxNumPublicIP int64 `url:"maxNumPublicIP,omitempty"`
 
 	// If true send emails when a user is granted access to resources
 	// Required: false
@@ -47,7 +47,7 @@ type CreateRequest struct {
 
 	// Limit (positive) or disable (0) GPU resources
 	// Required: false
-	GPUUnits uint64 `url:"gpu_units,omitempty"`
+	GPUUnits int64 `url:"gpu_units,omitempty"`
 }
 
 func (arq CreateRequest) validate() error {
@@ -69,7 +69,7 @@ func (a Account) Create(ctx context.Context, req CreateRequest) (uint64, error) 
 		return 0, err
 	}
 
-	url := "/cloudbroker/account/create"
+	url := "/cloudapi/account/create"
 
 	res, err := a.client.DecortApiCall(ctx, http.MethodPost, url, req)
 	if err != nil {