+++ pkg/cloudapi/rg/create.go
@@ -23,23 +23,23 @@ type CreateRequest struct {
 
 	// Max size of memory in MB
 	// Required: false
-	MaxMemoryCapacity uint64 `url:"maxMemoryCapacity,omitempty"`
+	MaxMemoryCapacity int64 `url:"maxMemoryCapacity,omitempty"`
 
 	// Max size of aggregated virtual disks in GB
 	// Required: false
-	MaxVDiskCapacity uint64 `url:"maxVDiskCapacity,omitempty"`
+	MaxVDiskCapacity int64 `url:"maxVDiskCapacity,omitempty"`
 
 	// Max number of CPU cores
 	// Required: false
-	MaxCPUCapacity uint64 `url:"maxCPUCapacity,omitempty"`
+	MaxCPUCapacity int64 `url:"maxCPUCapacity,omitempty"`
 
 	// Max sent/received network transfer peering
 	// Required: false
-	MaxNetworkPeerTransfer uint64 `url:"maxNetworkPeerTransfer,omitempty"`
+	MaxNetworkPeerTransfer int64 `url:"maxNetworkPeerTransfer,omitempty"`
 
 	// Max number of assigned public IPs
 	// Required: false
-	MaxNumPublicIP uint64 `url:"maxNumPublicIP,omitempty"`
+	MaxNumPublicIP int64 `url:"maxNumPublicIP,omitempty"`
 
 	// Username - owner of this resource group.
 	// Leave blank to set current user as owner